lambda/requirements-dev.txt
boto3-stubs[essential]
boto3-stubs[sqs]
